Improved user experience of deleting a film. A confirmation appears to make sure that the film the user wants to delete is the correct one

+++ deleteFilm.py
@@ -2,16 +2,29 @@ from connection import *
 from readFilms import *
 
 def deleteFilm():
-    filmID = int(input("""Enter the film ID of the film you wish to delete. 
-Print all the films first if you want to find the ID of the film you wish to delete
+    filmID = int(input("""Enter the film ID of the film you wish to delete.
     """))
 
-    cursor.execute(f"DELETE FROM tblFilms WHERE filmID = {filmID}")
-    connection.commit()
+    cursor.execute(f"SELECT * FROM tblFilms WHERE filmID = {filmID}")
 
-    read()
+    foundFilm = cursor.fetchone()
 
-    print(f"Film with ID {filmID} has been deleted")
+    if foundFilm:
+        confirmDelete = input(f"Your chosen film to delete is \"{foundFilm[1]}\" is this correct? Enter Y to confirm or anything else for No ").upper()
+        
+        if confirmDelete == "Y":
+            cursor.execute(f"DELETE FROM tblFilms WHERE filmID = {filmID}")
+            connection.commit()
+
+            read()
+
+            print(f"{foundFilm[1]} has been deleted")
+    
+        else:
+            print("The film was not deleted")
+    else:
+        print("That Film ID does not exist, no films have been deleted")
+
 
 if __name__ == "__main__":
     deleteFilm()
